+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Bug fixes
+- Fix issues with proposals to show software upgrade proposals #355
+
 ## 3.0.0
 
 ### Release Name: Antonio de Abreu